Introduce runtime parameter `update-no-snapshots`
This is `false` by default, in which the current snapshot is replaced by
a current update after each update operation, as before.

When set to `true`, this does not happen. This saves time, but means
that the update has no effect on queries (yet).

When changed from `true` to `false`, the current snapshot is replaced
according to the current state of the `DeltaTriples`. That is, if there
were updates while `update-no-snapshots` was `true`, these will now
become visible to queries.